requirements.txt
textual
pybluez; platform_system != "Windows"
pybluez-win10; platform_system == "Windows"
pyobjc-core; platform_system == "Darwin"
pyobjc-framework-CoreBluetooth; platform_system == "Darwin"
